Clear algorithm
Instrument class simplified, algorithm split up into smaller functions. Works with simulated instruments. Eval removed, much more readable :) The algorithm thread had to be modified in conjunction with the instrument classes. The big difference being that now the instrument class contains the actual instrument and sends the commands directly to it, as opposed to sending code-snippets to the main thread.